Split up integration tests to avoid OOM errors on CI
