+++ gradle.properties
@@ -1,3 +1,2 @@
-org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=1g -XX:MaxPermSize=2g
 android.useAndroidX=true
 kotlin.code.style=official
