+++ prefiller/src/test/kotlin/io/github/simonschiller/prefiller/integration/BaseIntegrationTest.kt
@@ -1,0 +1,11 @@
+package io.github.simonschiller.prefiller.integration
+
+import io.github.simonschiller.prefiller.testutil.ProjectExtension
+import org.junit.jupiter.api.extension.RegisterExtension
+
+abstract class BaseIntegrationTest {
+
+    @JvmField
+    @RegisterExtension
+    val project = ProjectExtension()
+}
