+++ prefiller/src/test/kotlin/io/github/simonschiller/prefiller/testutil/spec/BaseProjectSpec.kt
@@ -8,11 +8,7 @@ abstract class BaseProjectSpec : ProjectSpec {
 
     override fun getSettingsGradleContent(): String = "include(':module')"
     override fun getLocalPropertiesContent() = "sdk.dir=${getAndroidHome()}"
-
-    override fun getGradlePropertiesContent() = """
-        org.gradle.jvmargs=-Xmx4g -XX:MaxMetaspaceSize=1g -XX:MaxPermSize=2g
-        android.useAndroidX=true
-    """.trimIndent()
+    override fun getGradlePropertiesContent() = "android.useAndroidX=true"
 
     override fun getAndroidManifestContent() = """
         <manifest package="com.test" />
